requirements.txt
# Email работа (встроенные модули)
# imaplib - встроенный
# email - встроенный
# ssl - встроенный

# Работа с переменными окружения
python-dotenv==1.0.0

# Telegram бот
python-telegram-bot==20.7

# Логирование (расширенное)
loguru==0.7.2

# Работа с датами
python-dateutil==2.8.2

# JSON и файлы (встроенные)
# json - встроенный
# os - встроенный
# re - встроенный